requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
djangorestframework-simplejwt~=5.3.1
djangorestframework~=3.15.2
django-filter~=24.3
drf-yasg~=1.21.7
django-cors-headers~=4.4.0
